requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
Pillow
scikit-image
tensorflow
opencv-python
keras
streamlit
plotly
transformers
torch
accelerate
sentencepiece
protobuf
imbalanced-learn
